added function to format all numbers
